Push typedefs into `client::subcomponents` for tidier doc browsing.
