+++ src/client.rs
@@ -1,7 +1,9 @@
 //! Includes both `Client` and all of the RPC response types.
 #[macro_use]
 mod defapi;
+pub mod subcomponents;
 
+use self::subcomponents::{Consensus, NetworkUpgradeDesc, Softfork, ValuePool};
 use crate::{ResponseResult, ZecAmount};
 use reqwest;
 use serde::de::DeserializeOwned;
@@ -27,42 +29,6 @@ impl Client {
             reqcli: reqwest::Client::new(),
             idit: (0..),
         }
-    }
-
-    /// Construct a `Client` using the values of the environment variables `"ZCASHRPC_HOST"` and `"ZCASHRPC_AUTH"` as the arguments to `Client::new`.
-    pub fn from_env() -> Result<Client, std::env::VarError> {
-        use std::env::var;
-
-        let host = var("ZCASHRPC_HOST")?;
-        let auth = var("ZCASHRPC_AUTH")?;
-        Ok(Client::new(host, auth))
-    }
-
-    async fn make_request<R>(
-        &mut self,
-        method: &'static str,
-        args: Vec<serde_json::Value>,
-    ) -> ResponseResult<R>
-    where
-        R: DeserializeOwned,
-    {
-        use crate::{
-            envelope::{RequestEnvelope, ResponseEnvelope},
-            json,
-        };
-
-        let id = self.idit.next().unwrap();
-        let reqresp = self
-            .reqcli
-            .post(&self.url)
-            .header("Authorization", &self.auth)
-            .body(&RequestEnvelope::wrap(id, method, args))
-            .send()
-            .await?;
-        let text = reqresp.text().await?;
-        let respenv: ResponseEnvelope = json::parse_value(json::parse_string(text)?)?;
-        let resp = respenv.unwrap(id)?;
-        Ok(resp)
     }
 }
 
@@ -107,43 +73,31 @@ def_api_method! {
     }
 }
 
-#[derive(Debug, Deserialize, Serialize)]
-#[allow(non_snake_case)]
-pub struct ValuePool {
-    pub id: String,
-    pub monitored: bool,
-    pub chainValue: Option<ZecAmount>,
-    pub chainValueZat: Option<u64>,
-    pub valueDelta: Option<ZecAmount>,
-    pub valueDeltaZat: Option<i64>,
-}
+impl Client {
+    async fn make_request<R>(
+        &mut self,
+        method: &'static str,
+        args: Vec<serde_json::Value>,
+    ) -> ResponseResult<R>
+    where
+        R: DeserializeOwned,
+    {
+        use crate::{
+            envelope::{RequestEnvelope, ResponseEnvelope},
+            json,
+        };
 
-#[derive(Debug, Deserialize, Serialize)]
-pub struct Softfork {
-    pub id: String,
-    pub version: i64,
-    pub enforce: SoftforkMajorityDesc,
-    pub reject: SoftforkMajorityDesc,
-}
-
-#[derive(Debug, Deserialize, Serialize)]
-pub struct SoftforkMajorityDesc {
-    pub status: bool,
-    pub found: i64,
-    pub required: i64,
-    pub window: serde_json::Value, // FIXME
-}
-
-#[derive(Debug, Deserialize, Serialize)]
-pub struct NetworkUpgradeDesc {
-    pub name: String,
-    pub activationheight: u64,
-    pub status: String, // FIXME: enum-ify
-    pub info: String,
-}
-
-#[derive(Debug, Deserialize, Serialize)]
-pub struct Consensus {
-    pub chaintip: String,
-    pub nextblock: String,
+        let id = self.idit.next().unwrap();
+        let reqresp = self
+            .reqcli
+            .post(&self.url)
+            .header("Authorization", &self.auth)
+            .body(&RequestEnvelope::wrap(id, method, args))
+            .send()
+            .await?;
+        let text = reqresp.text().await?;
+        let respenv: ResponseEnvelope = json::parse_value(json::parse_string(text)?)?;
+        let resp = respenv.unwrap(id)?;
+        Ok(resp)
+    }
 }
